Refactor gym membership and plan interfaces for improved clarity and consistency
